Split test file into two, one for users and one for parcels
