+++ test/base.py
@@ -4,7 +4,9 @@ import unittest
 from app import app
 
 class BaseTestCase(unittest.TestCase):
-    """A base test case"""
+    """
+        A base test case
+    """
 
     def create_app(self):
         app.config.from_object('config.TestingConfig')
